Update Compatibility.lua
Commented ll. 79 & 80 because if you run the game with LUA errors enabled it turns annoying trying to open Dungeon Caches for example, since it's not a valid QuestID for the addon. I feel they should be commented and if someone is working on the addon they can simply remove those lines as comments